refactor: update subscriptionmanagement page to typescript. extract pause functions into hook diectory
